完善发布脚本

+++ action-eslint-jest/entrypoint.sh
@@ -3,18 +3,24 @@
 # avoid continuing when errors or undefined variables are present.
 set -eu
 
-# sync and update git submodules
-mkdir /root/.ssh/
-gpg -q --batch --yes --decrypt --passphrase="$LARGE_SECRET_PASSPHRASE" -o /root/.ssh/id_rsa docker_id_rsa.gpg
-chmod 400 /root/.ssh/id_rsa
+if [ -f docker_id_rsa.gpg ]; then
+    # sync and update git submodules
+    mkdir /root/.ssh/
+    gpg -q --batch --yes --decrypt --passphrase="$LARGE_SECRET_PASSPHRASE" -o /root/.ssh/id_rsa docker_id_rsa.gpg
+    chmod 400 /root/.ssh/id_rsa
 
-touch /root/.ssh/known_hosts
-ssh-keyscan github.com >> /root/.ssh/known_hosts
+    touch /root/.ssh/known_hosts
+    ssh-keyscan github.com >> /root/.ssh/known_hosts
+fi
 
-git submodule sync
-git submodule update --init
+if [ -f .gitmodules ]; then
+    git submodule sync
+    git submodule update --init
+fi
 
-rm -rf /root/.ssh
+if [ -d /root/.ssh ]; then
+    rm -rf /root/.ssh
+fi
 
 # 安装环境依赖#
 yarn install
